Fix float value

+++ src/callbacks.h
@@ -4,16 +4,15 @@
 #ifndef CALLBACKS_H
 #define CALLBACKS_H
 
-typedef struct _connection_state{
+typedef struct _connection_state {
     int port;
     bool connected;
     bool activated;
 } _connection_state;
 
 
-typedef struct _params
-{
-    char* host;
+typedef struct _params {
+    char *host;
     int port;
     int common_address;
     int server_number;
@@ -22,22 +21,20 @@ typedef struct _params
     int *ioas;
     int *types;
     int *types_gi;
-    int *vals;
-    int *deviations;
+    float *vals;
+    float *deviations;
 } _params;
 
 
-
-
-bool connectionRequestHandler(void* parameter, const char* ipAddress);
+bool connectionRequestHandler(void *parameter, const char *ipAddress);
 
 bool
-interrogationHandler(void* parameter, IMasterConnection connection, CS101_ASDU asdu, uint8_t qoi);
+interrogationHandler(void *parameter, IMasterConnection connection, CS101_ASDU asdu, uint8_t qoi);
 
 void
-connectionEventHandler(void* parameter, IMasterConnection con, CS104_PeerConnectionEvent event);
+connectionEventHandler(void *parameter, IMasterConnection con, CS104_PeerConnectionEvent event);
 
 bool
-clockSyncHandler (void* parameter, IMasterConnection connection, CS101_ASDU asdu, CP56Time2a newTime);
+clockSyncHandler(void *parameter, IMasterConnection connection, CS101_ASDU asdu, CP56Time2a newTime);
 
 #endif
